* Make sure to reset certain client variables serverside during map change * Clients now set their buffers as desynchronized during map change * Change errant size_t in nm_sectorposition_t to a uint32_t

+++ source/cs_main.h
@@ -823,7 +823,7 @@ typedef struct
 {
    int32_t message_type;
    uint32_t world_index;
-   size_t sector_number;
+   uint32_t sector_number;
    sector_position_t sector_position;
 } nm_sectorposition_t;
 
@@ -835,8 +835,8 @@ typedef struct
    int32_t message_type;
    uint32_t world_index;
    int32_t special_type; // [CG] map_special_e.
-   size_t line_number;
-   size_t sector_number;
+   uint32_t line_number;
+   uint32_t sector_number;
 } nm_specialspawned_t;
 
 // [CG] A map special's status is appended to the end of this message, the type
